+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 - Convert context to a list of contexts to allow for nested contexts in lexing.
+- Make Prague the default EVM version.
+- Update `revm` and `foundry` to the latest version.
 
 ## [1.1.7] - 2025-03-30
 - Throw error if the argument count for a macro call is not equal to the macro definition (Fixes: #49).